Registration and email
Worked on the registration system which sends an email to the new site
member.

Not 100% finished as the replacement system doesn't replace the
activation link key text correctly as it only displays the url of the
activation.

Need to make it so that an API front end is made to hide the backend
surface Controllers