Worked on Content and reintroduced the api system
Started work on the content and template system

Brought back the API page so that the backend system can be cleaner
this can be changed if the API page validations are removed and the
ajax gateway is then removed from the site which will then expose
the ajax backend without any code changes across the site

Added in dialog selections:
    Dialog
    Jumbotron
    Centered Dialog

Working on the Header and the Footer system to allow for custom
displays like templated current username output.

Added CKEditor to the backend for better content editing for fancier
items on the webpages.